requirements.txt
pydantic==2.8.2
pydantic-settings==2.3.4
fastapi==0.111.1
uvicorn==0.30.1
websockets==12.0
python-dotenv==1.0.1
google-genai==1.65.0
SpeechRecognition==3.10.4
pyttsx3==2.90
PyAudio==0.2.14
psutil==7.2.2
duckduckgo-search==7.5.0 
